Move edit settings display data functionality to SettingsUIMapper
COAND-860
